Add "make ghci-start", which runs a ghci that can load compiler + ghci
Summary:
Major credit to Csongor Kiss who wrote nearly all of settings.ghci. I've just
adding a few little tweaks that make it so that the loaded ghci can be run from
within ghci. There were three things that needed to be changed:

1. The compiler code requires NoImplicitPrelude, but GHCi's modules are incompatible with it. So, this adds the pragma to all GHCi modules that didn't have it, and adds imports of Prelude.

2. In order to run GHC within itself, a call of 'initGCStatistics' needed to be skipped. This uses CPP to skip it when -DGHC_LOADED_INTO_GHCI is set.

3. When running ":main", the inner GHC would attempt to find files like"lib/settings".  However, since "-B" is not provided, it would look in "inplace/lib/lib/settings".  I've added a workaround for this which is only active when "-DGHC_LOADED_INTO_GHCI" is set.

This also adds NoImplicitPrelude to template-haskell modules.  Loading
template-haskell + GHC into GHCi doesn't entirely work yet - it causes type
mismatches in TcSplice.  Even so, it can be quite useful to simultaneously
load template-haskell for iteration on that package.

It seemed prudent to have the workaround for the topdir location ("inplace/lib")
be a bit noisy that it's deviating from standard behavior, so I added putStrLns.
This revealed that 'findTopDir' is being called 3 times on GHC initialization.
So another part of this change is moving the calls of that function out of
'initLlvmConfig' and 'initSysTools', and having just one call in 'initGhcMonad'.
This way, top_dir can be computed once and shared with the initialization
functions.

Test Plan:
Manual for now. I have some thoughts on how to run the entire testsuite against
GHC-in-GHCi.

Reviewers: alpmestan, mpickering, goldfire, bgamari, erikd

Subscribers: rwbarton, thomie, carter

Differential Revision: https://phabricator.haskell.org/D4904